refactor: create a resuable useClusters hook and EmptyState component
